Exclude dependencies on a package basis Add a custom rollup plugin to resolve inferno dependencies to a relative path
