+++ config/rollup.config.js
@@ -2,30 +2,28 @@ import * as p from 'path';
 import * as fs from 'fs';
 import { rollup } from 'rollup';
 import buble from 'rollup-plugin-buble';
-import nodeResolve from 'rollup-plugin-node-resolve';
 import replace from 'rollup-plugin-replace';
 import uglify from 'rollup-plugin-uglify';
 import filesize from 'rollup-plugin-filesize';
 import pack from '../package.json';
 import commonjs from 'rollup-plugin-commonjs';
 import typescript from 'rollup-plugin-typescript';
-import { Bundles, getPackageJSON, outputFileSize } from './helpers';
+import { withNodeResolve, relativeModules, getPackageJSON, outputFileSize } from './rollup.helpers';
+import bundles from './rollup.bundles';
 
 const pkg = JSON.parse(fs.readFileSync('./package.json'));
-const external = Object.keys(pkg.peerDependencies || {}).concat(Object.keys(pkg.dependencies || {}));
+const dependencies = Object.keys(pkg.peerDependencies || {})
+													 .concat(Object.keys(pkg.dependencies || {}))
+													 .concat(Object.keys(pkg.devDependencies || {}));
 
-const plugins = [
+let plugins = [
 	typescript({
 		typescript: require('typescript')
 	}),
 	buble({
 		objectAssign: 'Object.assign'
 	}),
-	nodeResolve({
-		jsnext: true,
-		main: true,
-		skip: external
-	}),
+	relativeModules(),
 	commonjs({
 		include: 'node_modules/**',
 		exclude: ['node_modules/symbol-observable/**', '**/*.css']
@@ -66,92 +64,6 @@ if (process.env.NODE_ENV === 'production') {
 // Filesize plugin needs to be last to report correct filesizes when minified
 plugins.push(filesize());
 
-const bundles = new Bundles();
-
-bundles.add({
-	moduleGlobal: 'Inferno',
-	moduleName: 'inferno',
-	moduleEntry: 'packages/inferno/src/index.js',
-	path: 'packages/inferno/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoDOM',
-	moduleName: 'inferno-dom',
-	moduleEntry: 'packages/inferno-dom/src/index.js',
-	path: 'packages/inferno-dom/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoServer',
-	moduleName: 'inferno-server',
-	moduleEntry: 'packages/inferno-server/src/index.js',
-	path: 'packages/inferno-server/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoComponent',
-	moduleName: 'inferno-component',
-	moduleEntry: 'packages/inferno-component/src/index.js',
-	path: 'packages/inferno-component/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoTestUtils',
-	moduleName: 'inferno-test-utils',
-	moduleEntry: 'packages/inferno-test-utils/src/index.js',
-	path: 'packages/inferno-test-utils/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoCreateElement',
-	moduleName: 'inferno-create-element',
-	moduleEntry: 'packages/inferno-create-element/src/index.js',
-	path: 'packages/inferno-create-element/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoCompat',
-	moduleName: 'inferno-compat',
-	moduleEntry: 'packages/inferno-compat/src/index.js',
-	path: 'packages/inferno-compat/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoRouter',
-	moduleName: 'inferno-router',
-	moduleEntry: 'packages/inferno-router/src/index.js',
-	path: 'packages/inferno-router/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoCreateClass',
-	moduleName: 'inferno-create-class',
-	moduleEntry: 'packages/inferno-create-class/src/index.js',
-	path: 'packages/inferno-create-class/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoRedux',
-	moduleName: 'inferno-redux',
-	moduleEntry: 'packages/inferno-redux/src/index.js',
-	path: 'packages/inferno-redux/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoMobx',
-	moduleName: 'inferno-mobx',
-	moduleEntry: 'packages/inferno-mobx/src/index.js',
-	path: 'packages/inferno-mobx/'
-});
-
-bundles.add({
-	moduleGlobal: 'InfernoHyperscript',
-	moduleName: 'inferno-hyperscript',
-	moduleEntry: 'packages/inferno-hyperscript/src/index.js',
-	path: 'packages/inferno-hyperscript/'
-});
-
 function createBundle({ moduleGlobal, moduleName, moduleEntry }, path) {
 	const pack = getPackageJSON(moduleName, pkg);
 	const copyright =
@@ -174,8 +86,18 @@ function createBundle({ moduleGlobal, moduleName, moduleEntry }, path) {
 		sourceMap: false
 	};
 
-	return rollup({ entry, plugins }).then(({ write }) => write(bundleConfig)).catch(err => {
-		console.log(err);
+	const external = dependencies.concat(Object.keys(pack.dependencies || {}))
+
+	// Skip bundling dependencies of each package
+	plugins = withNodeResolve(plugins, {
+		module: true,
+		jsnext: true,
+		main: true,
+		skip: external
+	});
+
+	return rollup({ entry, plugins, external }).then(({ write }) => write(bundleConfig)).catch(err => {
+		console.log(err)
 	});
 }
