+++ src/mobx/connect.ts
@@ -1,6 +1,6 @@
 import invariant from 'invariant';
-import Component from '../component/es2015';
-import createClass from '../component/createClass';
+import Component from 'inferno-component';
+import createClass from 'inferno-create-class';
 import reactiveMixin from './reactiveMixin';
 import inject from './inject';
